repart: hardcode size for /usr partition to 5GB
This seems to be the only reliable way to make sure that sysupdate
has enough space to apply future updates. I picked 4GB because this
seems sensible, SteamOS uses 5GB for it's A/B slots and 4GB is probably
enough for the forseeable future (especially when using erofs with
compression).