profiles/compressed: remove mkosi-generated initrd
We want Bootable=yes (so mkosi installs sd-boot and builds/installs
a uki), but an unfortunate side effect of this option is that mkosi
generates a totally useless initrd in the output dir. Maybe one day it
won't be useless, but for now we don't want this polluting our manifest
or something.

Signed-off-by: Clayton Craft <craftyguy@postmarketos.org>